Extract caching logic into a hook
